refactor: extract Auto-TSTM state hook and move test harness out of __tests__
- Split useAutoTstmState from orchestration hook for CodeScene health
- Relocate shared test harness to src/hooks/testing to avoid empty Jest suite
