+++ src/hooks/__tests__/useAutoTstm.test.tsx
@@ -6,7 +6,7 @@ import {
   openPanelAndWaitForError,
   openPanelAndWaitForPreview,
   renderAutoTstm,
-} from './useAutoTstm.testHelpers';
+} from '../testing/autoTstmTestHarness';
 
 describe('useAutoTstm', () => {
   beforeEach(() => {
